perf(test): run isolated suites concurrently
Move VM-global attestation, semantic cache, and scheduler cases into serial companion modules while enabling async execution for isolated root and Viewer suites. Use ExUnit-managed temporary directories for Viewer filesystem isolation.

Verified with repeated max-concurrency stress runs, independent review, and mix precommit.
