+++ test/ptc_runner/kernel/command_initializer_test.exs
@@ -1,5 +1,5 @@
 defmodule PtcRunner.Kernel.CommandInitializerTest do
-  use ExUnit.Case, async: false
+  use ExUnit.Case, async: true
 
   alias PtcRunner.Kernel.CommandContract
   alias PtcRunner.Kernel.CommandEngine
