+++ test/ptc_runner/kernel/execution_session_owner_privacy_test.exs
@@ -1,5 +1,5 @@
 defmodule PtcRunner.Kernel.ExecutionSessionOwnerPrivacyTest do
-  use ExUnit.Case, async: false
+  use ExUnit.Case, async: true
 
   import ExUnit.CaptureLog
